+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 # Changelog
 
 
-## next version v0.0.3 - ???
+## version v0.3.0 - 20260505
 
 - fix typo #2
 - custom `gitonic` path -> file `~/.gitonic/thonnycontrib.cfg` 
 - changed to subprocess implementation
   - platform windows specfic popen and shlex handling
+- changed version number format to `0.3.x`
 -